Feature (Get Total Units)
What does this pr do?

Have a student able to fetch all the units he/she has registered for.
