+++ app/api/v1/views/subjects_view.py
@@ -69,6 +69,16 @@ def get_subjects_for_specific_user_by_admission(admission_no):
                                 "Subjects successfull retrieved")
 
 
+@portal_v1.route('/subjects/total/<string:admission_no>', methods=['GET'])
+@jwt_required
+def get_total_number_of_registered_units_by_admission(admission_no):
+    """Fetch the total number of registered units for a specific student."""
+    response = SubjectsModel().\
+        get_total_number_of_registered_units_by_admission(admission_no)
+    return Serializer.serialize(response, 200,
+                                "Total registered units")
+
+
 @portal_v1.route('/subjects/<string:admission_no>/<string:year>',
                  methods=['GET'])
 @jwt_required
